requirements.txt
pyttsx3
SpeechRecognition
wikipedia
pywhatkit
google-generativeai
python-dotenv
pyaudio